Fixed types
Fixed types in tryKick

+++ Robockey/PlayerLogic.cpp
@@ -225,9 +225,9 @@ void charge(){
 //Should be called in the move with puck
 void tryKick(){
 	Pose currentPose = getRobotPose();
-	int16_t dMax = ROBOT_RADIUS + 20;
-	int16_t dX = XMAX - currentPose.x;
-	int16_t dL = dX/cosb(currentPose.o);
+	uint8_t dMax = ROBOT_RADIUS + 20;
+	uint8_t dX = XMAX - currentPose.x;
+	float dL = dX/cosb(currentPose.o);
 	if(dL <= dMax){
 		int16_t dY = dL * sinb(currentPose.o);
 		int16_t goalY = currentPose.y + dY;
